cpu/requirements.txt
numpy==1.14.5
scikit-learn
scikit-image
flask
flask-cors
pygame
pyglet
opencv-python
opencv-contrib-python
pybind11
BriCA2
tensorflow
chainer
torch
torchvision
gym